AddStudent Enroll Button
Adding properly when all fields are filled!

But....when all fields are not filled, there is an error because it is selling int/float to an empty string which is not cool apparently.

I am trying to work out the logic in my head but I cant seem to get it right. I will eat lunch and then figure this out.